+++ picasso/gui/render.py
@@ -913,7 +913,7 @@ class PlotDialog(QtWidgets.QDialog):
         self.setWindowTitle("Structure")
         layout_grid = QtWidgets.QGridLayout(self)
 
-        self.figure = plt.figure()
+        self.figure = plt.figure(constrained_layout=True)
         self.canvas = FigureCanvas(self.figure)
         self.label = QtWidgets.QLabel()
 
@@ -1051,7 +1051,7 @@ class PlotDialogIso(QtWidgets.QDialog):
         self.setWindowTitle("Structure")
         layout_grid = QtWidgets.QGridLayout(self)
 
-        self.figure = plt.figure()
+        self.figure = plt.figure(constrained_layout=True)
         self.canvas = FigureCanvas(self.figure)
         self.label = QtWidgets.QLabel()
 
@@ -1283,7 +1283,7 @@ class ClsDlg3D(QtWidgets.QDialog):
         self.showMaximized()
         self.layout_grid = QtWidgets.QGridLayout(self)
 
-        self.figure = plt.figure()
+        self.figure = plt.figure(constrained_layout=True)
         self.canvas = FigureCanvas(self.figure)
         self.label = QtWidgets.QLabel()
 
@@ -1490,7 +1490,7 @@ class ClsDlg2D(QtWidgets.QDialog):
         self.setWindowTitle("Structure")
         self.layout_grid = QtWidgets.QGridLayout(self)
 
-        self.figure = plt.figure()
+        self.figure = plt.figure(constrained_layout=True)
         self.canvas = FigureCanvas(self.figure)
         self.label = QtWidgets.QLabel()
 
@@ -3053,7 +3053,7 @@ class DriftPlotWindow(QtWidgets.QTabWidget):
         icon = QtGui.QIcon(icon_path)
         self.setWindowIcon(icon)
         self.resize(1000, 500)
-        self.figure = plt.Figure()
+        self.figure = plt.Figure(constrained_layout=True)
         self.canvas = FigureCanvas(self.figure)
         vbox = QtWidgets.QVBoxLayout()
         self.setLayout(vbox)
@@ -3745,7 +3745,7 @@ class NenaPlotWindow(QtWidgets.QTabWidget):
         icon = QtGui.QIcon(icon_path)
         self.setWindowIcon(icon)
         self.resize(1000, 500)
-        self.figure = plt.Figure()
+        self.figure = plt.Figure(constrained_layout=True)
         self.canvas = FigureCanvas(self.figure)
         vbox = QtWidgets.QVBoxLayout()
         self.setLayout(vbox)
@@ -3778,7 +3778,7 @@ class FRCPlotWindow(QtWidgets.QTabWidget):
         icon = QtGui.QIcon(icon_path)
         self.setWindowIcon(icon)
         self.resize(1000, 500)
-        self.figure = plt.Figure()
+        self.figure = plt.Figure(constrained_layout=True)
         self.canvas = FigureCanvas(self.figure)
         vbox = QtWidgets.QVBoxLayout()
         self.setLayout(vbox)
@@ -5141,7 +5141,7 @@ class DisplaySettingsDialog(QtWidgets.QDialog):
         self.minimum_render.setRange(-999999, 999999)
         self.minimum_render.setSingleStep(5)
         self.minimum_render.setValue(0)
-        self.minimum_render.setDecimals(2)
+        self.minimum_render.setDecimals(5)
         self.minimum_render.setKeyboardTracking(False)
         self.minimum_render.setEnabled(False)
         self.minimum_render.valueChanged.connect(
@@ -5157,7 +5157,7 @@ class DisplaySettingsDialog(QtWidgets.QDialog):
         self.maximum_render.setRange(-999999, 999999)
         self.maximum_render.setSingleStep(5)
         self.maximum_render.setValue(100)
-        self.maximum_render.setDecimals(2)
+        self.maximum_render.setDecimals(5)
         self.maximum_render.setKeyboardTracking(False)
         self.maximum_render.setEnabled(False)
         self.maximum_render.valueChanged.connect(
@@ -5205,6 +5205,24 @@ class DisplaySettingsDialog(QtWidgets.QDialog):
             "Display the legend for the rendered property."
         )
         render_grid.addWidget(self.show_legend, 5, 1)
+
+        fw, fh, dpi = (2, 1, 150)
+        self.figure_prop = plt.figure(
+            figsize=(fw, fh), dpi=dpi, constrained_layout=True
+        )
+        self.ax_prop = self.figure_prop.add_subplot(111)
+        # adjust the size of ticks and labels
+        self.ax_prop.tick_params(
+            axis="both", which="both", labelsize=4, length=2, width=0.5, pad=1
+        )
+
+        self.canvas_prop = FigureCanvas(self.figure_prop)
+        self.canvas_prop.setSizePolicy(
+            QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Expanding
+        )
+        self.canvas_prop.setMinimumSize(QtCore.QSize(fw * dpi, fh * dpi))
+        render_grid.addWidget(self.canvas_prop, 6, 0, 6, 2)
+
         self.show_legend.setEnabled(False)
         self.show_legend.setAutoDefault(False)
         self.show_legend.clicked.connect(self.window.view.show_legend)
@@ -5287,6 +5305,41 @@ class DisplaySettingsDialog(QtWidgets.QDialog):
         """Update scene if dynamic display pixel size is checked."""
         if state:
             self.window.view.update_scene()
+
+    def update_histogram(self) -> None:
+        """Update histogram of the rendered property based on the
+        current min/max values and colors."""
+        self.ax_prop.cla()
+
+        # array of values for the rendered property
+        data = self.window.view.locs[0][self.parameter.currentText()].values
+        # other parameters
+        min_val = self.minimum_render.value()
+        max_val = self.maximum_render.value()
+        if max_val == min_val:
+            max_val += 1e-6  # avoid zero division
+        n_colors = self.color_step.value()
+        colors = get_render_properties_colors(
+            n_colors, self.colormap_prop.currentText()
+        )
+
+        # plot
+        n_bins = lib.calculate_optimal_bins(data, max_n_bins=100)
+        counts, bins, patches = self.ax_prop.hist(
+            data, bins=n_bins, range=(min_val, max_val)
+        )
+        for patch, bin_left in zip(patches, bins):
+            color_idx = int(
+                (n_colors - 1) * (bin_left - min_val) / (max_val - min_val)
+            )
+            color_idx = np.clip(color_idx, 0, n_colors - 1)
+            patch.set_facecolor(colors[color_idx])
+        self.ax_prop.set_xlim(min_val, max_val)
+        self.ax_prop.set_xlabel("")
+        self.ax_prop.set_ylabel("")
+        self.ax_prop.set_yticks([])
+        self.ax_prop.xaxis.offsetText.set_fontsize(4)  # scientific notation
+        self.canvas_prop.draw()
 
     def update_scene(self, *args, **kwargs) -> None:
         """Update scene with cache."""
@@ -5892,16 +5945,13 @@ class View(QtWidgets.QLabel):
             return
         locs = lib.ensure_sanity(locs, info)
 
-        # update pixelsize
-        for element in info:
-            expression = r"Picasso( v\d+\.\d+\.\d+)? Localize"
-            if re.search(
-                expression, element.get("Generated by", ""), re.IGNORECASE
-            ):
-                if "Pixelsize" in element:
-                    self.window.display_settings_dlg.pixelsize.setValue(
-                        element["Pixelsize"]
-                    )
+        # update pixelsize (credits to Boyd Peters #602)
+        pixelsize = lib.get_from_metadata(
+            info,
+            "Pixelsize",
+            default=self.window.display_settings_dlg.pixelsize.value(),
+        )
+        self.window.display_settings_dlg.pixelsize.setValue(pixelsize)
 
         # append loaded data
         self.locs.append(locs)
@@ -6188,7 +6238,7 @@ class View(QtWidgets.QLabel):
 
             # Plot shift
             if display:
-                fig1 = plt.figure(figsize=(8, 8))
+                fig1 = plt.figure(figsize=(8, 8), constrained_layout=True)
                 plt.suptitle("Shift")
                 plt.subplot(1, 1, 1)
                 plt.plot(shift_x, "o-", label="x shift")
@@ -8541,7 +8591,9 @@ class View(QtWidgets.QLabel):
                     params["t0"] = time.time()
                     i = 0
                     while i < len(self._picks):
-                        fig = plt.figure(figsize=(5, 5))
+                        fig = plt.figure(
+                            figsize=(5, 5), constrained_layout=True
+                        )
                         fig.canvas.manager.set_window_title(
                             "Scatterplot of Pick"
                         )
@@ -8624,7 +8676,9 @@ class View(QtWidgets.QLabel):
                     i = 0
                     while i < len(self._picks):
                         pick = self._picks[i]
-                        fig = plt.figure(figsize=(5, 5))
+                        fig = plt.figure(
+                            figsize=(5, 5), constrained_layout=True
+                        )
                         fig.canvas.manager.set_window_title(
                             "Scatterplot of Pick"
                         )
@@ -9014,7 +9068,7 @@ class View(QtWidgets.QLabel):
                 progress.close()
 
                 # plot histogram with n_locs in picks
-                fig = plt.figure()
+                fig = plt.figure(constrained_layout=True)
                 fig.canvas.manager.set_window_title("Localizations in Picks")
                 ax = fig.add_subplot(111)
                 ax.set_title("Localizations in Picks ")
@@ -10120,7 +10174,7 @@ class View(QtWidgets.QLabel):
             self.window.display_settings_dlg.colormap_prop.currentText(),
         )
 
-        fig1 = plt.figure(figsize=(5, 1))
+        fig1 = plt.figure(figsize=(5, 1), constrained_layout=True)
 
         ax1 = fig1.add_subplot(111, aspect="equal")
 
@@ -10206,6 +10260,7 @@ class View(QtWidgets.QLabel):
 
             self.x_locs = x_locs
             self.window.display_settings_dlg.show_legend.setEnabled(True)
+            self.window.display_settings_dlg.update_histogram()
         else:
             self.x_render_state = False
         self.update_scene()
@@ -10257,6 +10312,10 @@ class View(QtWidgets.QLabel):
 
         self.window.display_settings_dlg.render_check.setEnabled(True)
         self.window.display_settings_dlg.render_check.setCheckState(False)
+
+        # clean up the histogram plot
+        self.window.display_settings_dlg.ax_prop.cla()
+        self.window.display_settings_dlg.canvas_prop.draw()
 
         self.activate_render_property()
 
@@ -10970,8 +11029,6 @@ class View(QtWidgets.QLabel):
             True if optimally adjust contrast. Default is False.
         use_cache : bool, optional
             True if use stored image. Default is False.
-        cache : bool, optional
-            True if save image. Default is False.
         picks_only : bool, optional
             True if only picks and points are to be rendered. Default is
             False.
@@ -11012,8 +11069,6 @@ class View(QtWidgets.QLabel):
             True if optimally adjust contrast. Default is False.
         use_cache : bool, optional
             True if use stored image. Default is False.
-        cache : bool, optional
-            True if save image. Default is False.
         picks_only : bool, optional
             True if only picks and points are to be rendered. Default is
             False.